Remove set_value_optimized_out.
Improve the api for marking values optimized out, the api is now similar in
style to that used for marking values unavailable.  The only (possible)
issue is that this commit reverts this commit (by Pedro):
	http://sourceware.org/ml/gdb-patches/2013-07/msg00178.html
however, later patches should restore the world to a happy state.  I
believe that the current code is functional, and was only rejected on the
grounds that /if/ in the future we added support for values being both lazy
AND optimized out then the above would break.

gdb/ChangeLog

	* ada-lang.c (coerce_unspec_val_to_type): Switch to
	mark_value_bytes_optimized_out, and only mark non-lazy values as
	optimized out.
	* dwarf2loc.c (read_pieced_value): Switch to use
	mark_value_bytes_optimized_out, or mark_value_bits_optimized_out.
	(write_pieced_value): Switch to use
	mark_value_bytes_optimized_out.
	* findvar.c (read_frame_register_value): Switch to use
	mark_value_bytes_optimized_out.
	(value_from_register): Likewise.
	* value.c (allocate_optimized_out_value, value_fetch_lazy): Switch
	to use mark_value_bytes_optimized_out.
	(value_optimized_out_const, set_value_optimized_out): Remove.
	(mark_value_bytes_optimized_out, mark_value_bits_optimized_out):
	New function definition.
	* value.h (value_optimized_out_const, set_value_optimized_out):
	Remove.
	(mark_value_bytes_optimized_out, mark_value_bits_optimized_out):
	Add new declaration.