Add handling of R_MRK3_AUTO16 relocation.
This relocation is generated by LLVM as it is unable to determine in the
case of jump tables/constant tables whether a symbol being referenced is
in a code section or data section and therefore whether it should be
a word or data pointer. In this case, the linker decides based on the
section (if it has the code flag set) and shifts accordingly.

bfd:
	* elf32-mrk3.c (elf_mrk3_howto_table): Add R_MRK3_AUTO16 relocation.
	(mrk3_final_link_relocate): Add function to automatically determine
	whether to bitshift depending on whether symbol is in code space.
	(mrk3_elf_relocate_section): Added to utilize
	mrk3_final_link_relocate.

include:
	* elf/mrk3.h: Add R_MRK3_AUTO16 relocation.