gdb-disassembler: Handle more instruction formats from simlator.
There are some layouts of instructions from the simulator that the
current gdb code does not handle correctly, this leads to corrupted
output when disassembling.

This patch restructures the parser within gdb to (hopefully) make it
handle these missing cases better.

The main problem that seems to trip up gdb is that the current simulator
is inconsistent as to when it places a '0x' prefix onto hex values.

ChangeLog:

	* mrk3-tdep.c (skip_whitespace): New function.
	(parse_opcode_hex): New function.
	(parse_opcode_and_args): New function.
	(mrk3_fancy_print_insn): Replace use of sscanf with use of new
	parser functions.