When unwinding the PC off the stack, only load 16-bits.
The CALL instruction only writes 16-bits of the PC onto the stack.  As
we model the PC as a 32-bit register, then gdb will, by default, try to
fetch 32-bits from the stack, this not only fetches too many bits, thus
accessing, possibly invalid memory[1] but leaves the resulting PC value
with possibly non-zero bits in the upper 16-bits.

[1] OK, it's not likely that fetching from the stack will result in an
access to invalid memory, but we should really only access memory that
we know is valid.

ChangeLog:

	* mrk3-tdep.c (struct mrk3_frame_cache): New structure.
	(mrk3_analyze_prologue): Change parameter type. Initialise
	saved_regs set.  Write register data into saved_regs set instead
	of using a trad_frame_cache.
	(mrk3_alloc_frame_cache): New function.
	(mrk3_frame_cache): Update return type.  Allocate mrk3_frame_cache
	rather than a trad_frame_cache.
	(mrk3_frame_base_address): Extract the base from mrk3_frame_cache.
	(mrk3_frame_prev_register): Use saved_regs from mrk3_frame_cache
	to unwind registers.  Add special case handling for
	MRK3_PC_REGNUM.