gdb: Improve unwinding of stack pointer.
We had two members of the mrk3_frame_cache for tracking the stack
pointer on entry to the frame 'prev_sp' and 'base'.  The 'prev_sp'
member was set correctly but never used, while 'base' was set
incorrectly.

Given that other targets (based on random sampling) use 'base', I've
gone with that too.

In order to unwind the stack pointer we should use the computed value,
to achieve this the computed value needs to be stored using
trad_frame_set_value.

gdb/ChangeLog:

	* mrk3-tdep.c (struct mrk3_frame_cache): Remove prev_sp, this is
	the same as base.
	(mrk3_analyze_prologue): No longer set prev_sp member of the cache
	structure.  Set the base memory of the cache struture to th
	previous stack pointer (prev_sp), and set the value of the
	previous stack pointer register in the traditional frame cache.