gdb: Decode 4-bit constants in prologue analysis.
The 4-bit short constants are actually encoded from a look-up table, the
value within the instruction encoding is /not/ the actual value of the
constant.  This look up table mechanism was not implemented, and so in
some cases we were using the wrong short immediate as the stack offset.

gdb/ChangeLog:

	* mrk3-tdep.c (mrk3_decode_const4): New function.
	(mrk3_analyze_prologue): Use mrk3_decode_const4.