ld: Changes to how code relocations are scaled.
We have some relocations that can reference either code or data.  In
these cases the choice on whether the value being patched in should be a
byte address, or word address is made in `mrk3_final_link_relocate`.

However, previously, we scaled the relocation value and then called
_bfd_final_link_relocate to finish the relocation.  The problem with
this is that for pc-relative relocations the scaled value is then used
to compute an offset from an unscaled pc, with the results that the
value we attempt to patch in would be wrong, and in most cases would
overflow.

To fix this issue I've moved the contents of _bfd_final_link_relocate
into mrk3_final_link_relocate, and inserted the scaling after the
pc-relative adjustment has been performed.  We can then call
_bfd_relocate_contents to insert the value into the instruction.

For non pc-relative relocations this should make no difference.

ChangeLog:

	* elf32-mrk3.c (mrk3_final_link_relocate): Extend comment.  Add
	copy of code from _bfd_final_link_relocate into here.  Remove
	adjustment of the r_addend, and call _bfd_relocate_contents to
	finish the relocation.