gdb: Only access a single register to find the frame base.
For MRK32 data pointers are, in general, 32-bits.  However, the stack
pointer is 16-bits.  This does mean that the stack is restricted to
living in the lower address region.

In the DWARF debugging information, the address size is given as
4-bytes, to accommodate the general 32-bit data pointer.  However, the
frame base for a function is specified as simple the SP register.

The problem then, is that gdb will create a value of type data pointer,
which is 32-bits large, and then fills this by reading from register
SP.  And here's the problem, if a register is smaller than the required
size, gdb will automatically wrap to the "next" register (next here is
defined simply by gdbs internal register ordering).  The result is that
the frame base we extracted is being polluted with the contents of the
register that is adjacent to the stack pointer.

Given that DWARF should not depend on this wrap over feature (a value
split over multiple register will be explicitly defined as being in
specific multiple register using the DW_OP_piece mechanism) I have taken
the approach of modifying the read address from register mechanism used
by DWARF such that we first read the contents of /only/ the register in
question, and then we cast the contents to the required address type.
In this way, the 16-bit stack pointer read from the SP register, should
be zero extended to be a 32-bit address.

gdb/ChangeLog:

	* findvar.c (value_from_register): Access the register as it's
	native type, then cast to the desired address type.