opcodes: Fix configure.in and configure files.
In the opcodes directory, the configure file had been edited directly
rather than updating the configure.in file.  Fix this by back porting
the bfd_mrk3_arch change to configure.in and regenerate.

The regeneration adds back some code to turn on -Werror by default,
however, as we configure binutils with --disable-werror this should not
cause any problems.

opcodes/ChangeLog:

	* configure.in: Add bfd_mrk3_arch.
	* configure: Regenerate.