MRK3/binutils: Use MRK3 register names in DWARF dump.
When dumping the DWARF from an ELF, use the real MRK3 register names as
much as possible.

This patch adds an outline table for mapping DWARF register number to
MRK3 register names.  I've filled in the bits I understand right now,
those I've not filled in will fall back to using the standard
notation (that is "r" followed by a number).

binutils/ChangeLog:

	* dwarf.c (dwarf_regnames_mrk3): New table.
	(init_dwarf_regnames_mrk3): New function.
	(init_dwarf_regnames): Add EM_MRK3 case.
	* dwarf.h (init_dwarf_regnames_mrk3): Declare.
	* objdump.c (dump_dwarf): Add bfd_arch_mrk3 case.