MRK3/gdb: Turn on the register convert mechanism.
On MRK3 we have a problem.  Addresses within gdb are 32-bit, this
includes space for the target pointer (24-bit) and some space for an
address space identifier (8-bits) [1].

However, there are many pointers on the target that are only ever
16-bits in length, the stack pointer falls into this group, also some
structures are known to have been placed within the 16-bit address
range, and can, therefore be accessed using 16-bit pointers.

Now, in the DWARF information, we will say, for example, the stack
pointer is stored in register R7, or the pointer to variable foo is in
register R4.  When this happens gdb creates a value of type
TYPE_CODE_PTR which is backed by the register in question, and then we
fetch the value from the register.

The problem is that the size of the value is larger than the size of the
register as gdb thinks that all pointers are 32-bits, and the registers
in question are just 16-bits.  What happens then is that gdb will
actually fetch 16-bits from the requested register, and the remaining
16-bits from the "next" register[2].

I think that the real solution to this issue might be to teach gdb that
we have pointers of different sizes.  However, that feels like a huge
piece of work, and so this is my temporary solution.

I've turned on the register conversion mechanism for MRK3.  Using this
mechanism, I can catch places where we access a register in order to
read a pointer, and control how the access takes place.  In this case I
read the 16-bit register, and then zero extend to make a 16-bit
pointer.  When writing to a register (so the user is setting a pointer
variable), I check that the most significant 16-bits are clear, then
write out the lower 16-bits.

I have not fully thought through how this will work with longer pointer
types, but then I'm not clear what the DWARF will look like in those
cases.  I figure I'll come back to that when I have some solid examples
to work with.

[1] This in itself might be a future issue as in some cases the target
needs to support /real/ pointers up to 32-bits long.  However, that's a
problem for another day.

[2] Next here just means that gdb increments its internal register
number, and we access whatever register that corresponds to.

gdb/ChangeLog:

	* mrk3-tdep.c (mrk3_convert_register_p): New function.
	(mrk3_register_to_value): New function.
	(mrk3_value_to_register): New function.
	(mrk3_gdbarch_init): Register new gdbarch functions.