ld: Linker-maps, don't access undefined symbols in PROVIDE.
This follows on from a821504850e163bcabfce07bcd61c6bc0f6ce374, and
reflects what was actually committed to the upstream tree.

There is no need to use a look up to check if the PROVIDE is actually
being used or not, the expression class will tell us directly, the class
will be either etree_provide (unused) or etree_provided (used).

ld/ChangeLog:

	* ldlang.c (print_assignment): Remove symbol lookup before
	evaluating expression.