gdb: Add early gdbarch hook into frame unwind process.
Currently the frame unwind mechanism has a hard coded check for a PC
value of zero, which causes the frame unwind to stop.  This is fine for
most architectures, but in some cases this check is not appropriate.

Two examples of where this check might not be correct are:

  1. An architecture that allows code to be mapped at address
  zero (small embeded for example), in which case, stopping at address
  zero might be incorrect.

  2. Harvard architecture machines, for example AVR, which use
  additional bits within the address to represent different address
  spaces.  In this case an address of zero might not be seen (due to the
  address space identifier) but if the non identifier part of the
  address is zero then the unwind should still stop.

This patch makes the zero PC check a feature of gdbarch, allowing
different architectures to override it as appropriate.  To allow for
things other than just a zero-pc check, the new gdbarch function returns
the debug string explaining why the unwind is stopping (if the unwind
should stop) or NULL (if the unwind should continue).

gdb/ChangeLog:

	* frame.c: Add gdbarch.h include.
	(get_prev_frame): Call gdbarch_unwind_stop_at_frame_p instead of
	checking for zero pc.
	(default_unwind_stop_at_frame_p): New function.
	* frame.h (default_unwind_stop_at_frame_p): Declare.
	* gdbarch.c: Add frame.h include.
	(struct gdbarch): Add unwind_stop_at_frame_p member.
	(startup_gdbarch): Initialise unwind_stop_at_frame_p.
	(gdbarch_alloc): Initialise unwind_stop_at_frame_p.
	(gdbarch_dump): Dump unwind_stop_at_frame_p.
	(gdbarch_unwind_stop_at_frame_p): New function.
	(set_gdbarch_unwind_stop_at_frame_p): New function.
	* gdbarch.h (struct frame): Declare.
	(gdbarch_unwind_stop_at_frame_p_ftype): New type.
	(gdbarch_unwind_stop_at_frame_p): Declare.
	(set_gdbarch_unwind_stop_at_frame_p): Declare.