gdb: Add MRK3 unwind zero pc check.
As we start to carry around the address space identifier in the PC value
the zero pc check in the frame unwind code will no longer work for
us (the address space identifier will make the pc appear as always non
zero).

This patch adds an MRK3 specific function that first removes the address
space identification bits, and then checks for a zero pc.

gdb/ChangeLog:

	* mrk3-tdep.c (mrk3_unwind_stop_at_frame_p): New function.
	(mrk3_gdbarch_init): Register mrk3_unwind_stop_at_frame_p.