gdb: Tighten logic in register conversion mechanism.
We use the register conversion mechanism to zero extend pointers (which
gdb thinks are 32-bit) that are held in 16-bit registers.  MRK3 does
have some 32-bit registers, so we only need to convert if fetching from
a 16-bit register.  Performing the conversion in all cases is not wrong,
just unneeded (and possibly confusing to developers) so this patch
tightens the check, a conversion is only needed if we are fetching a
pointer that is different in size to the register size.

gdb/ChangeLog:

	* mrk3-tdep.c (mrk3_convert_register_p): Change param name.
	Tighten logic to include a check of the type lengths.