gdb: Don't remove the address space bits from pc values.
If we strip the address space identification bits from the PC too
quickly then we will be unable to determine which address space we are
referring too.

When reading and unwinding the PC we should maintain the address space
identification bits.

The remove address bits hook is (currently) used when loading addresses
from the DWARF, it is critical that address loaded from DWARF maintain
their address space identification bits, and so, we should ensure that
we don't register the remove address bits hook.  If we do remove the
address space identification bits (as we currently do) then we are
unable to map any address found in DWARF back to a symbol.

ChangeLog:

	* mrk3-tdep.c (mrk3_addr_bits_remove): Delete.
	(mrk3_read_pc): Don't call mrk3_addr_bits_remove.
	(mrk3_breakpoint_from_pc): Don't mess with the PC passed in.
	(mrk3_unwind_pc): Don't call mrk3_addr_bits_remove.
	(mrk3_gdbarch_init): Don't register mrk3_addr_bits_remove.