gdb: Observer frame_cache_cleared, not normal_stop.
We were using the normal_stop observer to clear the cached memory space
identifier.  The problem is that gdb prints the current location before
calling the normal stop observers, the result is that when gdb stops it
reuses the not yet invalidated address space identifier, which may no
longer be valid.

The frame_cache_cleared observer triggers whenever gdb thinks that the
frames might be invalid, this is an excellent time to clear out cached
state such as the memory space identifier.

gdb/ChangeLog:

	* mrk3-tdep.c (mrk3_handle_normal_stop) Rename to ...
	(mrk3_handle_frame_cache_cleared) ... this, and update log message
	inside function.
	(_initialize_mrk3_tdep): Install frame_cache_cleared observer, not
	normal_stop.