bfd: Mask out address space id before adjusting relocations.
When computing the value that is to be patched into a relocation we must
first remove the address space id bits.  This is required because
leaving the id bits in place when, for example, scaling an address, or
computing a pc-relative value, will result in the address space id
becoming corrupted.

ChangeLog:

	* elf32-mrk3.c (MRK3_GET_MEMORY_SPACE_ID): New macro.
	(MRK3_GET_ADDRESS_LOCATION): New macro.
	(MRK3_BUILD_ADDRESS): New macro.
	(mrk3come_final_link_relocate): Avoid adjusting addresses contains
	memory space id bits.