bfd/ld: Add R_MRK3_CONST4 and R_MRK3_CALL14 relocations.
Two new relocation R_MRK3_CONST4 and R_MRK3_CALL14, neither of which are
generated by the assembler.  Both of these will be generated as a result
of linker relaxation which will be added in later commits.

The R_MRK3_CONST4 relocation allows the common, 4-bit encoded constants
to be fixed up in a relocation.  If the assembler knows the value of a
symbol, it will generate a fully patched instruction using either a
4-bit, or 16-bit constant as appropriate.  If the assembler does not
know the value of a symbol then it has no way of knowing that the value
will be one that can be encoded using the 4-bit constants and so will
always fall back to a 16-bit relocation.

With linker relaxation thought the linker can check the value to be
patched in and reduce the instruction from a 32-bit instruction taking a
16-bit immediate, to a 16-bit instruction taking a 4-bit immediate.

The R_MRK3_CALL14 relocation is used in a shorter version of the call
instruction, the full size call is a 32-bit instruction with a 16-bit
address encoded within, the short call is a 16-bit instruction with a
14-bit address encoded within.

bfd/ChangeLog:

	* elf32-mrk3.c (elf_mrk3_howto_table): Add R_MRK3_CONST4 and
	R_MRK3_CALL14 entries.
	(mrk3_final_link_relocate_const4): New function.
	(mrk3_final_link_relocate): Add support for new relocation types.

include/ChangeLog:

	* elf/mrk3.h: Add R_MRK3_CONST4 and R_MRK3_CALL14 relocation
	numbers.