bfd: Fix bitsize of R_MRK3_HIGH16 relocation.
The bitsize of a relocation is used during overflow checking, and
represents the number of bits of the relocation value that are encoded
within the instruction.  If more than this number of bits are
significant within the relocation value then overflow has occurred.

The R_MRK3_HIGH16 relocation incorrectly had a bitsize of 32, as a
result bfd was failing to spot overflows within these relocations.

bfd/ChangeLog:

	* elf32-mrk3 (elf_mrk3_howto_table): Change bitsize of
	R_MRK3_HIGH16 to 16 bits.