mrk3: Add initial linker relaxation support.
This commit adds initial support for linker relaxation.  Relaxation is
off by default, but will be turned on by passing the -relax flag to the
linker.

The following relaxations are performed:

  - Relax 16-bit call instructions to 14-bit call instructions.
  - Relax 16-bit immediate instructions to 4-bit immediate instructions
    where possible.  Generally the assembler will take care of these for
    us, but, if the immediate is a symbol then the assembler can't relax
    the instruction, but the linker can.  This does crop up sometimes.
  - Relax 16-bit pc-relative branches to 8-bit pc-relative branches.

There are still some problems with the current approach to linker
relaxation that should be addressed in future commits.

The problem is that when relaxing CALL16 to CALL14 it is not uncommon to
find a situation where a CALL14 is moved to an earlier 14-bit page as a
result of /other/ optimisations performed by linker relaxation.

When this change of page occurs the newly created CALL14 will no longer
link, the result is that linker relaxation produces a binary that does
not link, when the original unrelaxed binary would have linked.

This link failure is not acceptable, ideally the linker relaxation
mechanism should be able to spot that a particular relaxation has become
invalid and should "undo" the relaxation, in this way we will reach a
binary that is as relaxed as we can manage, without making the binary
invalid.

It is worth noticing thought that the above issue means that if linker
relaxation does complete then the resulting binary is valid.  For now
the message is, give linker relaxation a try, if the link completes then
your binary is safe, if linker relaxation results in a link error then
this is unfortunate, for now you'll need to skip linker relaxation.

bfd/ChangeLog:

	* elf64-mrk3: Add include of bfd_stdint.h.
	(retrieve_internal_relocs): New function.
	(pin_internal_relocs): New function.
	(release_internal_relocs): New function.
	(retrieve_contents): New function.
	(pin_contents): New function.
	(release_contents): New function.
	(retrieve_local_syms): New function.
	(release_local_syms): New function.
	(relax_log): New function.
	(elf64_mrk3_relax_delete_bytes): New function.
	(is_16bit_call_instruction): New function.
	(is_16bit_immediate_instruction): New function.
	(elf64_mrk3_relax_section): New function.
	(bfd_elf64_bfd_relax_section): Define.