bfd: Cast to signed vma before scaling to word addresses.
When patching pc relative relocations to code symbols we scale the bfd
byte address to a mrk3 word address.  As pc relative relocation can be
positive or negative it is critical that the most significant bit on
negative values be maintained, otherwise a negative value can appear to
become a large positive value after the scaling, which will then cause
the relocation to be reported as an overflow.

bfd/ChangeLog:

	* elf32-mrk3.c (mrk3_final_link_relocate): Cast bfd_vma to
	bfd_signed_vma before shifting.