gdb: Don't initialise architecture specific gdbarch data.
The gdbarch structure has a field for holding a pointer to an
architecture specific structure.  We don't currently use this mechanism,
but do define a dummy structure which we allocate and pass in at
initialisation time.

There was a todo in the code asking if it was possible to /not/
initialise the architecture specific data, it is, other targets manage
it, so we can too.

Allocating, and passing around unused data structures only makes the
code more confusing to understand, it's not clear at a casual glance
that these structure are unused.  Better to delete them now, and just
add them back if we ever want them again in the future.

gdb/ChangeLog:

	* mrk3-tdep.c (struct gdbarch_tdep): Delete.
	(mrk3_gdbarch_init): Don't initialise architecture specific tdep
	field.
	(mrk3_dump_tdep): Delete.
	(_initialize_mrk3_tdep): Don't pass mrk3_dump_tdep.