gdb: New type for memory space id.
Add a new type for the MRK3 memory space id instead of using uint32_t,
this is in preparation to moving to 64-bit ELF format.

The new type is just CORE_ADDR, as that is how the memory space id is
used; it's merged into and out of MRK3 addresses.

As a side effect of this change I had to clean up how we print the
memory space id values, they now have a print format string defined at
the top of the file, which can be easily changed when we move to 64-bit
ELF.

While pushing the new type into the code I've made a few whitespace
cleanups in one function (mrk3_get_memspace), but nothing massive.

In mrk3_pseudo_register_write I've updated an internal type from
uint32_t to int, this means that the internal variable is now the same
type as the function return type.

gdb/ChangeLog:

	* mrk3-tdep.c (MRK3_ELF_ADDRESS_SIZE): Define.
	(MRK3_MEM_SPACE_BYTE): Define, use, then undefine.
	(MRK3_MEM_SPACE_MASK): Define using MRK3_MEM_SPACE_BYTE.
	(MRK3_MEM_SPACE_NONE): Likewise.
	(MRK3_MEM_SPACE_SYS): Likewise.
	(MRK3_MEM_SPACE_APP1): Likewise.
	(MRK3_MEM_SPACE_APP2): Likewise.
	(MRK3_MEM_SPACE_SSYS): Likewise.
	(MRK3_MEM_TYPE_MASK): Likewise.
	(MRK3_MEM_TYPE_DATA): Likewise.
	(MRK3_MEM_TYPE_CODE): Likewise.
	(MRK3_MEM_MASK): Likewise.
        (PRIxMEM_SPACE): Define.
	(mrk3_get_memspace): Update types to CORE_ADDR.  Some whitespace
	cleanups.  Parse string from gdbserver as long long.  Use new
	print format string.
	(mrk3_is_ssys_memspace): Use CORE_ADDR type.
	(mrk3_is_sys_memspace): Likewise.
        (mrk3_pseudo_register_read): Use new print format string.
        (mrk3_pseudo_register_write): Likewise.
	(mrk3_dwarf2_reg_to_regnum): Use int within function to match
	return type.
	(mrk3_p2a): Use CORE_ADDR type.
	(mrk3_addr_bits_add): Likewise.
	(mrk3_address_to_pointer): Likewise.
	(mrk3_read_pc): Likewise.
	(mrk3_analyze_prologue): Likewise.
	(mrk3_frame_this_id): Likewise.