linker-relaxation: Allow CALL16 to be relaxed.
In commit e5fe89153915b9efba87f0a651a0b8a817a36692 (initial linker
relaxation support) a mistake was made.  The code committed was a
version that pre-dated the use of the R_MRK3_CALL16 relocation.

This patch fixes that mistake.  The R_MRK3_CALL16 relocation is now
detected, and relaxed correctly.

The issue discussed in the commit e5fe891539... is still true, when a
CALL16 is relaxed to a CALL14, if the CALL14 moves onto a different
14-bit page then the link will fail.  This needs to be solved, in the
future, by adding support to check and undo relaxations that were too
aggressive.  For now the only answer is for the user to try the link
again with relaxation turned off.

bfd/ChangeLog:

	* elf64-mrk3.c (is_16bit_immediate_instruction): Renamed to...
	(is_relaxable_16bit_immediate_instruction): ... this.
	(elf64_mrk3_relax_section): Split the CALL16 handling out from the
	HIGH16 handling.