bfd/ld: Linker relaxation, check and undo support.
While performing linker relaxation, if a CALL16 is converted to a
CALL14, then it is possible that the source and destination of the
CALL14 end up on different 14-bit pages (due to code motion as a result
of /other/ relaxation).

When this happens the linker will fail while trying to patch the
CALL14 (with relocation overflow error).  This is not ideal as the only
option at that point is to not turn on linker relaxation for this link,
loosing any benefit.

Furthermore, as a project changes it is possible that a build that
yesterday worked fine with linker relaxation will suddenly stop
linking.  This makes managing the build process, and project development
difficult.

The solution, introduced here, is to add a new phase to linker
relaxation.  After the relaxation pass, we now perform a check pass.  In
the check pass the linker re-checks the relocations that were created by
linker relaxation, and if a relocation is found to no longer be
resolvable, the relocation will be reverted to its pre-relaxed state.

The solution proposed here is not the best possible solution.  The best
possible solution would, after performing any undo, look for new places
that could be relaxed.  However, the benefit likely obtained from this
extra relaxation is minimal, and the effort required to perform the
additional relaxation and checking passes would be quite large, so, for
now at least, I have let the process as a single relax phase followed by
a single checking phase.

bfd/ChangeLog:

	* elf64-mrk3.c (struct mrk3_relax_info): New struct.
	(struct elf_mrk3_section_data): New struct.
	(get_mrk3_relax_info): New function.
	(init_mrk3_relax_info): New function.
	(elf_mrk3_new_section_hook): New function.
	(pin_contents): Add assert.
	(retrieve_local_syms): Add header comment.
	(release_local_syms): Delete.
	(elf64_mrk3_relax_delete_bytes): Renamed to ...
	(mrk3_elf_relax_delete_bytes): ... this.  Improve comments
	throughout.  Record that a section was relaxed.
	(mrk3_elf_relax_insert_bytes): New function.
	(is_14bit_call_instruction): New function.
	(mrk3_relax_section_filter): New function.
	(mrk3_relax_relocation_filter): New function.
	(mrk3_relax_handle_relocation): New function.
	(mrk3_check_section_filter): New function.
	(mrk3_check_relocation_filter): New function.
	(mrk3_check_handle_relocation): New function.
	(typedef section_filter_ftype): New type.
	(typedef reloc_filter_ftype): New type.
	(typedef handle_reloc_ftype): New type.
	(struct mrk3_relaxation_hooks): New struct.
	(mrk3_relax_hooks): New global.
	(mrk3_check_hooks): New global.
	(mrk3_elf_relax_section_worker): New function, takes most of the
	code that used to be in elf64_mrk3_relax_section, but now uses
	hooks to eithr relax or check as required.
	(elf64_mrk3_relax_section): Renamed to ...
	(mrk3_elf_relax_section): ... this.  Now uses
	mrk3_elf_relax_section_worker to perform either relaxation or
	checking.
	(bfd_elf64_bfd_relax_section): Update function name.
	(bfd_elf64_new_section_hook): Define.

ld/ChangeLog:

	* emultempl/mrk3.em (gld${EMULATION_NAME}_after_parse): New
	function.
	(LDEMUL_AFTER_PARSE): Points to new function.