gdb: nxpload support.
This change makes the nxpload command pass all sections to the target.

In order to pass all sections, and associated data (lma, vma, flags,
section name) the `target_rcmd' mechanism is used to pass a new custom
command (called 'nxpload') to the target.

The new 'nxpload' command carries all of the section information that
the target requires, except for the content.  Once the 'nxpload' command
has been sent to the target, the section content is sent using the
standard 'X' packets.

When gdbserver receives the 'nxpload' command it will create a temporary
buffer into which the section contents will be gathered.  The section
contents then arrive in a series of 'X' packets.  Once all of the
section contents have been gathered by gdbserver the section can be
processed, maybe loading the contents into target memory, or configuring
the target.

A new flag 'nxpload-logging' is available for set/show use within gdb.
This control the logging output from nxpload.  This is not debugging
output, just helpful logging information.

When 'nxpload-logging' is set to 1 (the default) then the trace output
from 'nxpload' will be similar to the output from a standard 'load'
command.

When 'nxpload-logging' is set to 2 then the load progress of all
sections will be logged, this is more output than the user would get
from a standard 'load' command.

For completeness 'nxpload-logging' can be set to 0 to reduce the amount
of logging seen.

gdb/ChangeLog:

	* mrk3-tdep.c (nxpload_logging): New global.
	(struct nxpload_progress_section_data): New vma and flags members.
	(mrk3_load_transfer_section): New function.
	(mrk3_load_section_callback): Use mrk3_load_transfer_section
	function.  Handle loading zero sized sections.  Store vma and
	flags.
	(mrk3_nxpload_progress): Give user control over logging.  Make the
	logging MI friendly.
	(mrk3_load_sort_write_requests): New function.
	(mrk3_nxpload): Sort sections, then pass them to the target.
	(_initialize_mrk3_tdep): Setup control of nxpload_logging global.