bfd/ld: Honour the no-relax section flag.
The 'R' section flag marks a section as being non-relaxable.  However,
this flag was not being tested within the linker relaxation code.  The
result was that marking a section within the assembler with the 'R' flag
would not prevent the section being relaxed.

This patch fixes this oversight.

bfd/ChangeLog:

	* elf64-mrk3.c (mrk3_relax_section_filter): Don't relax sections
	marked with the no-relax section flag (SHF_MRK3_NON_RELAX).