gdb: Use bfd size as another cache identification criteria.
Within gdb open bfd objects are reused where possible if an attempt is
made to reopen a file that is already being debugged.  To spot if the on
disc file has changed gdb currently examines the mtime of the file and
compares it to the mtime of the open bfd in the cache.

A problem exists when the on disc file is being rapidly regenerated, as
happens, for example, with automated testing.  In some cases the file is
generated so quickly that the mtime appears not to change, while the on
disc file has changed.

This patch extends the bfd cache to also hold the file size of the bfd,
something which can be cheaply acquired, gdb can then check the mtime
and the file size.  This is still not going to catch every on disc
change, but does increase the probability that a change will be spotted.

gdb/ChangeLog:

	* gdb_bfd.c (struct gdb_bfd_data): Add size field.
	(struct gdb_bfd_cache_search): Likewise.
	(eq_bfd): Compare the size fields.
	(gdb_bfd_open): Initialise the size field.
	(gdb_bfd_ref): Likewise.
	(gdb_bfd_unref): Likewise.