Fix LONG register access, and PSW flags access.
This commit makes fixes to the pseudo register read / write code, and so
fixes issues when accessing the LONG register forms, for example R4LONG.

In the same area this commit also fixes some mistakes with the bit field
masks used to extract the various flag bits (carry, overflow, zero,
negative) from the PSW register.  After this commit the various flags
will all be extracted as either 0 or 1.

gdb/ChangeLog:

	* mrk3-tdep.c (mrk3_pseudo_register_read): Use MRK3_R4_REGNUM for
	clarity.  Fixes for big endian and little endian register access
	for the LONG register forms.  Fixes for the bit mask used when
	accessing fields from the PSW.
	(mrk3_pseudo_register_write): Likewise, but reversed to handle the
	write mode.